Fix the address used after establishing the connection
Especially when using docker networks, one might not want to specify the
exact ip address but a hostname which then is resolved to an ip address.
In order that the lookups later (eg for pow) succeed, the address added
to eg the toBeProven set must be the same we see when receiving a
packet. Therefore, now hz opens the connection and then notifies the
strategy which connections are established but now uses the resolved ip
address at this place.

Fixes https://gitlab.lrz.de/netintum/teaching/p2psec_projects_2024/Gossip-3/-/merge_requests/42#note_3326685